+++ lib/doorkeeper/config.rb
@@ -194,8 +194,7 @@ and that your `initialize_models!` method doesn't raise any errors.\n
     option :realm,                         default: 'Doorkeeper'
     option :wildcard_redirect_uri,         default: false
     option :force_ssl_in_redirect_uri,     default: !Rails.env.development?
-    option :grant_flows,
-           default: %w(authorization_code implicit password client_credentials)
+    option :grant_flows,                   default: %w(authorization_code client_credentials)
 
     attr_reader :reuse_access_token