+++ spec/requests/flows/implicit_grant_errors_spec.rb
@@ -3,6 +3,7 @@ require 'spec_helper_integration'
 feature 'Implicit Grant Flow Errors' do
   background do
     config_is_set(:authenticate_resource_owner) { User.first || redirect_to('/sign_in') }
+    config_is_set(:grant_flows, ["implicit"])
     client_exists
     create_resource_owner
     sign_in